requirements.txt
black==23.9.0
fairlearn==0.9.0
h5py==3.9.0
imageio==2.31.1
matplotlib==3.7.2
natsort==8.4.0
numpy==1.24.3
opencv-python==4.8.0.76
pandas==2.0.3
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.13.2
torchsummary==1.5.1
tqdm==4.66.1
